# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Production work reporting API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Operator {
  id: number;
  name: string;
  employeeId: string;
  isAdmin: boolean;
  isLineleader: boolean;
  isModerator: boolean;
  email?: string | null;
  createdAt: string;
}

export interface CreateOperatorBody {
  name: string;
  employeeId: string;
  isAdmin?: boolean;
  isLineleader?: boolean;
  isModerator?: boolean;
  email?: string | null;
}

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  /** @maxLength 10 */
  revision?: string | null;
  createdAt: string;
}

export interface Step {
  id: number;
  productId: number;
  stepNumber: number;
  subStepLabel?: string | null;
  name: string;
  description?: string | null;
  standardTimeMinutes: number;
}

export interface ProductWithSteps {
  id: number;
  name: string;
  description?: string | null;
  /** @maxLength 10 */
  revision?: string | null;
  createdAt: string;
  steps: Step[];
}

export interface CreateProductBody {
  name: string;
  description?: string | null;
  /** @maxLength 10 */
  revision?: string | null;
}

export interface CreateStepBody {
  stepNumber: number;
  subStepLabel?: string | null;
  name: string;
  description?: string | null;
  standardTimeMinutes: number;
}

export interface WorkReport {
  id: number;
  operatorId: number;
  productId: number;
  stepId: number;
  timeWorkedMinutes: number;
  quantityCompleted: number;
  operatorCount?: number | null;
  reportDate: string;
  notes?: string | null;
  createdAt: string;
  operator?: Operator;
  product?: Product;
  step?: Step;
}

export interface CreateReportBody {
  operatorId: number;
  productId: number;
  stepId: number;
  timeWorkedMinutes: number;
  quantityCompleted: number;
  operatorCount?: number | null;
  reportDate: string;
  notes?: string | null;
}

export interface DashboardSummary {
  totalOperators: number;
  totalProducts: number;
  totalReports: number;
  totalQuantityCompleted: number;
  totalTimeMinutes: number;
  recentReports: WorkReport[];
}

export interface OperatorStats {
  operatorId: number;
  operatorName: string;
  employeeId: string;
  totalReports: number;
  totalQuantityCompleted: number;
  totalTimeMinutes: number;
}

export interface ProductStats {
  productId: number;
  productName: string;
  totalReports: number;
  totalQuantityCompleted: number;
  totalTimeMinutes: number;
  stepCount: number;
}

export interface WeeklyPlan {
  id: number;
  productId: number;
  weekStart: string;
  plannedQuantity: number;
  createdAt: string;
  updatedAt: string;
  product?: Product;
}

export interface CreateWeeklyPlanBody {
  productId: number;
  weekStart: string;
  plannedQuantity: number;
}

export interface WeeklyProgress {
  productId: number;
  productName: string;
  plannedQuantity: number;
  completedQuantity: number;
  remainingQuantity: number;
  percentageComplete: number;
  planId?: number | null;
}

export type ListReportsParams = {
  operatorId?: number;
  productId?: number;
  date?: string;
};

export type ListWeeklyPlansParams = {
  /**
   * ISO date of the Monday (YYYY-MM-DD)
   */
  weekStart: string;
};

export type GetWeeklyProgressParams = {
  /**
   * ISO date of the Monday (YYYY-MM-DD)
   */
  weekStart: string;
};